# InnoSync-Mobile-App/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateProjectRequest:
    properties:
      description:
        type: string
      team_size:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  models.CreateProjectRoleRequest:
    properties:
      expertise_level:
        $ref: '#/definitions/models.ExpertiseLevelEnum'
      role_name:
        type: string
      technologies:
        items:
          type: string
        type: array
    required:
    - role_name
    type: object
...
  models.ExpertiseLevelEnum:
    enum:
    - ENTRY
    - JUNIOR
    - MID
    - SENIOR
    - RESEARCHER
    type: string
    x-enum-varnames:
    - ExpertiseLevelEntry
    - ExpertiseLevelJunior
    - ExpertiseLevelMid
    - ExpertiseLevelSenior
    - ExpertiseLevelResearcher
...
  models.Project:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      recruiter_id:
        type: integer
      roles:
        description: Related data
        items:
          $ref: '#/definitions/models.ProjectRole'
        type: array
      team_size:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.ProjectResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      roles:
        items:
          $ref: '#/definitions/models.ProjectRole'
        type: array
      team_size:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.ProjectRole:
    properties:
      expertise_level:
        $ref: '#/definitions/models.ExpertiseLevelEnum'
      id:
        type: integer
      is_open:
        type: boolean
      project:
        allOf:
        - $ref: '#/definitions/models.Project'
        description: Related data
      project_id:
        type: integer
      role_name:
        type: string
      technologies:
        type: string
    type: object
  models.ProjectRoleResponse:
    properties:
      expertise_level:
        $ref: '#/definitions/models.ExpertiseLevelEnum'
      id:
        type: integer
      is_open:
        type: boolean
      project_id:
        type: integer
      project_title:
        type: string
      role_name:
        type: string
      technologies:
        items:
          type: string
        type: array
    type: object
...
  models.UpdateProjectRequest:
    properties:
      description:
        type: string
      team_size:
        type: integer
      title:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: RESTful API for InnoSync Mobile Application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: InnoSync API
  version: "1.0"
paths:
...
  /projects:
    post:
      consumes:
      - application/json
      description: Create a new project
      parameters:
      - description: Project creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - projects
  /projects/{id}:
    delete:
      description: Delete a project (soft delete)
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - projects
    get:
      description: Get a specific project by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get project
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - projects
  /projects/{id}/roles:
    get:
      description: Get all roles for a specific project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProjectRoleResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get project roles
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Create a new role for a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project role creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateProjectRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProjectRoleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create project role
      tags:
      - projects
  /projects/{id}/roles/{roleId}:
    get:
      description: Get a specific project role by ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProjectRoleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get project role
      tags:
      - projects
  /projects/me:
    get:
      description: Get all projects created by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProjectResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user projects
      tags:
      - projects
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"